+++ src/ui.rs
@@ -452,7 +452,7 @@ impl Ui {
         frame.render_widget(gauge, chunks[chunk_start + 2]);
     }
 
-    fn video_percent_done_default(stage: VideoStage) -> f64 {
+    fn video_percent_done_default(stage: &VideoStage) -> f64 {
         match stage {
             // When a video is already present before starting the app,
             // then this video will be finished without `video.percent_done`