+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Tighten code style to pedantic level.
+- Signal child process shutdown to global shutdown process, rather than polling child stages.
 
 ## [0.11.1] - 2023-10-21